Fix rating input and header formatting issues on page 5 courses
Critical fixes:
- Changed rating input from (int) to (float) to preserve half-star ratings (2.5, 3.5, etc.)
- Removed "Write a Review" button from Jackson Country Club header to match Island Pointe format

Fixed rating input casting in:
- Jackson Country Club
- Memphis Country Club
- Mirimichi Golf Course
- Moccasin Bend Golf Course

Now all page 5 courses properly save and display half-star ratings and have consistent header formatting.

+++ courses/mirimichi-golf-course.php
@@ -47,7 +47,7 @@ $is_logged_in = SecureSession::isLoggedIn();
 
 // Handle comment submission
 if ($_SERVER['REQUEST_METHOD'] === 'POST' && $is_logged_in) {
-    $rating = (int)$_POST['rating'];
+    $rating = (float)$_POST['rating'];
     $comment_text = trim($_POST['comment_text']);
     $user_id = $_SESSION['user_id'];
